+++ src/main/java/com/pekar/callofcompanions/controllers/VanillaTeleportController.java
@@ -2,6 +2,7 @@ package com.pekar.callofcompanions.controllers;
 
 import com.mojang.logging.LogUtils;
 import com.pekar.callofcompanions.entity.EntityRegistry;
+import com.pekar.callofcompanions.data.CompanionEntry;
 import com.pekar.callofcompanions.scheduler.CompanionEntryScheduler;
 import com.pekar.callofcompanions.scheduler.CompanionEntryTask;
 import net.minecraft.core.BlockPos;
@@ -11,6 +12,7 @@ import org.slf4j.Logger;
 class VanillaTeleportController extends LoadedAnimalSummonController
 {
     private static final Logger LOGGER = LogUtils.getLogger();
+    private static final int MAX_ANIMAL_DISTANCE_TO_AVOID_TELEPORTING = 12;
 
     protected VanillaTeleportController(SummonAnimalContext context)
     {
@@ -20,7 +22,7 @@ class VanillaTeleportController extends LoadedAnimalSummonController
     @Override
     public void run(BlockPos teleportPos)
     {
-        int delay = level.getRandom().nextIntBetweenInclusive(applyDelayFactor(10), applyDelayFactor(100));
+        int delay = playerLevel.getRandom().nextIntBetweenInclusive(applyDelayFactor(10), applyDelayFactor(100));
         var task = new CompanionEntryTask(
                 delay,
                 companionEntry,
@@ -28,7 +30,7 @@ class VanillaTeleportController extends LoadedAnimalSummonController
                 (ticks, entry) -> {
                     if (ticks == 29)
                     {
-                        showAnimalTeleportParticles(level, animal);
+                        showAnimalTeleportParticles(playerLevel, animal);
                     }
                     else if (ticks == 9)
                     {
@@ -38,40 +40,45 @@ class VanillaTeleportController extends LoadedAnimalSummonController
                     return false;
                 },
                 entry -> {
-                    LOGGER.debug("Vanilla teleport completing: companionType={}, companionId={}", entry.type(), entry.uuid());
-                    if (animal.distanceToSqr(player) > 12 * 12)
-                    {
-                        var teleported = tryTeleportAnimalTo(level, animal.getUUID(), teleportPos, false);
-                        if (teleported)
-                        {
-                            playTeleportSound(level, animal);
-                            showAnimalTeleportParticles(level, animal);
-
-                            if (teleportListener != null)
-                                teleportListener.onTeleport(TeleportType.VANILLA_TELEPORT);
-                        }
-                        else
-                        {
-                            var name = CallCrystalHelper.buildAnimalName(entry.type(), entry.name());
-                            player.sendSystemMessage(Component.translatable("message.callofcompanions.cant_teleport", name), true);
-                            LOGGER.debug("Far teleport failed: companion couldn't find a safe place to teleport, companionType={}, companionId={}", entry.type(), entry.uuid());
-                        }
-                    }
-                    recreateAnimal(level, animal, animal.getX(), animal.getY(), animal.getZ());
-                    setGoal(animal, player);
-                    CallCrystalHelper.updateCompanionPos(level, companionData, entry);
+                    moveAnimalTo(teleportPos, entry);
                 },
                 entry -> {
-                    playAnimalNotRespondSound(level, teleportPos.below());
-                    showAnimalNotRespondParticles(level, teleportPos.below());
+                    LOGGER.debug("Vanilla teleport cancelled: companionType={}, companionId={}", companionEntry.type(), companionEntry.uuid());
+                    playAnimalNotRespondSound(playerLevel, teleportPos.below());
+                    showAnimalNotRespondParticles(playerLevel, teleportPos.below());
                     if (animal.getType().is(EntityRegistry.ANIMALS_CAN_TELEPORT_TO_PLAYER) && animal.distanceToSqr(player) < 10 * 10)
                     {
-                        recreateAnimal(level, animal, animal.getX(), animal.getY(), animal.getZ());
+                        recreateAnimal(playerLevel, animal, animal.getX(), animal.getY(), animal.getZ());
                     }
-                    LOGGER.debug("Vanilla teleport cancelled: companionType={}, companionId={}", companionEntry.type(), companionEntry.uuid());
                 }
         );
         CompanionEntryScheduler.UPDATE_POS_TASKS.add(task);
         LOGGER.debug("Vanilla teleport scheduled: companionType={}, companionId={}, delayTicks={}", companionEntry.type(), companionEntry.uuid(), delay);
+    }
+
+    private void moveAnimalTo(BlockPos teleportPos, CompanionEntry entry)
+    {
+        LOGGER.debug("Vanilla teleport completing: companionType={}, companionId={}", entry.type(), entry.uuid());
+        if (animal.distanceToSqr(player) > MAX_ANIMAL_DISTANCE_TO_AVOID_TELEPORTING * MAX_ANIMAL_DISTANCE_TO_AVOID_TELEPORTING)
+        {
+            var teleported = tryTeleportAnimalTo(playerLevel, animal.getUUID(), teleportPos, entry.dimension(), false);
+            if (teleported)
+            {
+                playTeleportSound(playerLevel, animal);
+                showAnimalTeleportParticles(playerLevel, animal);
+
+                if (teleportListener != null)
+                    teleportListener.onTeleport(TeleportType.VANILLA_TELEPORT);
+            }
+            else
+            {
+                var name = CallCrystalHelper.buildAnimalName(entry.type(), entry.name());
+                player.sendSystemMessage(Component.translatable("message.callofcompanions.cant_teleport", name), true);
+                LOGGER.debug("Vanilla teleport failed: companion couldn't find a safe place to teleport, companionType={}, companionId={}", entry.type(), entry.uuid());
+            }
+        }
+        recreateAnimal(playerLevel, animal, animal.getX(), animal.getY(), animal.getZ());
+        setGoal(animal, player);
+        CallCrystalHelper.updateCompanionPos(playerLevel, companionData, entry);
     }
 }